requirements.txt
requests==2.31.0
selenium==4.18.1
webdriver-manager==4.0.1
scdl>=2.0.0
python-dotenv>=0.19.0 